libs/langgraph-elasticsearch/pyproject.toml
[tool.poetry]
name = "langgraph-elasticsearch"
version = "0.1.0"
description = "Library with a Elasticsearch implementation of LangGraph checkpoint saver."
authors = []
readme = "README.md"
license = "MIT"
packages = [{ include = "langgraph" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
langchain = "^0.3.20"
langchain-elasticsearch = "^0.3.2"
langchain-openai = "^0.3.8"
langgraph = ">=0.2.76"
langgraph-checkpoint = "^2.0.19"
langchain-community = "^0.3.19"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.3"
pydantic-settings = "^2.8.1"
fitz = "^0.0.1.dev2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
testcontainers = "^4.9.2"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest-watcher]
now = true
delay = 0.1
patterns = ["*.py"]

[tool.poetry.scripts]
start = "langgraph.examples:main"


# [tool.poetry.group.test.dependencies.tests]
# path = "tests"
# develop = true
[tool.pytest.ini_options]

# --strict-markers will raise errors on unknown marks.
# https://docs.pytest.org/en/7.1.x/how-to/mark.html#raising-errors-on-unknown-marks
#
# https://docs.pytest.org/en/7.1.x/reference/reference.html
# --strict-config       any warnings encountered while parsing the `pytest`
#                       section of the configuration file raise errors.
addopts = "--strict-markers --strict-config --durations=5 -vv"
asyncio_mode = "auto"

[tool.vivoai]
title = "VivoAI API"
root_path = "/api"
docs_url = "/docs"
redoc_url = "/redoc"